Separated algorithms, basic trading functions, and backtesting code into individual files for readability. Will make testing a variety of algorithms in parallel more accessible.
